requirements.txt
# Azure Key Vault Rotation Demo - Python Dependencies
# This file contains all Python dependencies for the entire project

# Core Azure dependencies
azure-identity==1.24.0
azure-keyvault-secrets==4.10.0
azure-keyvault-keys==4.11.0
azure-keyvault-certificates==4.10.0

# Web dashboard dependencies
flask==3.0.3
python-dotenv==1.0.1

# Development and CLI dependencies
click>=8.0.0
colorama>=0.4.0
tabulate>=0.9.0

# Optional: Add these if you want enhanced functionality
# requests>=2.31.0
# pydantic>=2.0.0
